A lot of restructuring to allow for more abstraction
Improvements to address spaces as well, moved ACPI to MCore so all
architectures are allowed to use it. Some code cleanup as well,
improvements to Routing Tables handling code so it handles multiple
entries per device

+++ kernel/System/Syscalls.c
@@ -644,7 +644,7 @@ int ScIpcRead(uint8_t *MessageContainer)
 		return -1;
 
 	/* Read rest of message */
-	if (BytesRead != MsgBase->Length) 
+	if ((size_t)BytesRead != MsgBase->Length) 
 	{
 		/* Calculate remainder and increase pointer */
 		int Remainder = (MsgBase->Length - BytesRead);
